feat: misc improvements to infra deploy scripts
