+++ typescript/infra/.gitignore
@@ -3,6 +3,7 @@ tmp.ts
 dist/
 .env*
 cache/
+deployment-plan.yaml
 test/outputs
 config/environments/test/core/
 config/environments/test/igp/
