+++ typescript/sdk/src/hook/HyperlaneHookDeployer.ts
@@ -9,7 +9,12 @@ import {
   ProtocolFee,
   StaticAggregationHook__factory,
 } from '@hyperlane-xyz/core';
-import { Address, addressToBytes32, rootLogger } from '@hyperlane-xyz/utils';
+import {
+  Address,
+  addressToBytes32,
+  deepEquals,
+  rootLogger,
+} from '@hyperlane-xyz/utils';
 
 import { HyperlaneContracts } from '../contracts/types.js';
 import { CoreAddresses } from '../core/contracts.js';
@@ -166,7 +171,13 @@ export class HyperlaneHookDeployer extends HyperlaneDeployer<
       aggregatedHooks.push(subhooks[hookConfig.type].address);
       hooks = { ...hooks, ...subhooks };
     }
-    this.logger.debug(`Deploying aggregation hook of ${config.hooks}`);
+
+    this.logger.debug(
+      { aggregationHook: config.hooks },
+      `Deploying aggregation hook of type ${config.hooks.map((h) =>
+        typeof h === 'string' ? h : h.type,
+      )}...`,
+    );
     const address = await this.ismFactory.deployStaticAddressSet(
       chain,
       this.ismFactory.getContracts(chain).staticAggregationHookFactory,
@@ -309,7 +320,7 @@ export class HyperlaneHookDeployer extends HyperlaneDeployer<
       this.logger.debug(`Deploying routing hook for ${dest}`);
       const destDomain = this.multiProvider.getDomainId(dest);
 
-      if (prevHookConfig && prevHookAddress && prevHookConfig === hookConfig) {
+      if (deepEquals(prevHookConfig, hookConfig) && prevHookAddress) {
         this.logger.debug(`Reusing hook ${prevHookAddress} for ${dest}`);
         routingConfigs.push({
           destination: destDomain,
